conformance/buffimgs/cl_context_createImage.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLContext createImage conformance test.</title>
    <link rel="stylesheet" href="../../resources/js-test-style.css"/>
    <script src="../../resources/js-test-pre.js"></script>
    <script src="../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("WebCLContext createImage test case.");
    debug("(CLenum memFlags, WebCLImageDescriptor descriptor, optional ArrayBuffer hostPtr)");
    var wtu = WebCLTestUtils;
    var webCLContext = wtu.createContext();
    var webCLPlatforms = wtu.getPlatforms();
    var webCLDevices = wtu.getDevices(webCLPlatforms[0], webcl.DEVICE_TYPE_CPU);
    var imageWidth = 100;
    var imageHeight = 100;

    var data = new Float32Array(SIZE);
    count = SIZE;
    for (var i = 0; i < count; i++) {
        data[i] = Math.random();
    }

    var webCLImageDescriptorRead = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, imageWidth, imageHeight);
    var webCLImageDescriptorWrite = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_WRITE_ONLY, imageWidth, imageHeight);
    var webCLImageDescriptorReadWrite = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_WRITE, imageWidth, imageHeight);

    // Testing for MEM_READ_ONLY images.
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorRead)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorRead, data)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorWrite)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorWrite, data)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorReadWrite)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_ONLY, webCLImageDescriptorReadWrite, data)", WebCLImage);


    // Testing for MEM_WRITE_ONLY images.
    shouldBeType("webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorRead)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorRead, data)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorWrite)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorWrite, data)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite, data)", WebCLImage);

    // Testing for MEM_READ_WRITE images.
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorRead)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorRead, data)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorWrite)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorWrite, data)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite)", WebCLImage);
    shouldBeType("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite, data)", WebCLImage);

    // Testing for Exceptions.
    shouldThrowExceptionName("webCLContext.createImage(invalid_object, webCLImageDescriptorObject);", "INVALID_VALUE");

    // Testing for invalid values of Image width.
    var invalidImageWidth = 0;
    webCLImageDescriptor = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, invalidImageWidth, imageHeight);
    shouldThrowExceptionName("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptor);", "INVALID_IMAGE_SIZE");
    var imageWidthMax = webCLDevices[0].getInfo(webcl.DEVICE_IMAGE2D_MAX_WIDTH);
    invalidImageWidth = imageWidthMax + 1;
    webCLImageDescriptor = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, invalidImageWidth, imageHeight);
    shouldThrowExceptionName("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptor);", "INVALID_IMAGE_SIZE");

    // Testing for invalid values of Image height.
    var invalidImageHeight = 0;
    webCLImageDescriptor = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, imageWidth, invalidImageHeight);
    shouldThrowExceptionName("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptor);", "INVALID_IMAGE_SIZE");
    var imageHeightMax = webCLDevices[0].getInfo(webcl.DEVICE_IMAGE2D_MAX_HEIGHT);
    invalidImageHeight = imageHeightMax + 1;
    webCLImageDescriptor = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, imageWidth, invalidImageHeight);
    shouldThrowExceptionName("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptor);", "INVALID_IMAGE_SIZE");

    // Testing for invalid value of channelOrder.
    webCLImageDescriptor = webCLContext.getSupportedImageFormats(webcl.MEM_READ_ONLY);
    var webCLImageDescriptorObject = {width:imageWidth, height:imageHeight, channelOrder:invalid_object, channelType:webCLImageDescriptor[0].channelType};
    shouldThrowExceptionName("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorObject);", "INVALID_IMAGE_FORMAT_DESCRIPTOR");

    // Testing for invalid value of channelType.
    webCLImageDescriptor = webCLContext.getSupportedImageFormats(webcl.MEM_READ_ONLY);
    var webCLImageDescriptorObject = {width:imageWidth, height:imageHeight, channelOrder:webCLImageDescriptor[0].channelOrder, channelType:invalid_object};
    shouldThrowExceptionName("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptorObject);", "INVALID_IMAGE_FORMAT_DESCRIPTOR");

    //Testing for INVALID_HOST_POINTER.
    webCLImageDescriptor = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_ONLY, imageWidth, imageHeight);
    var min = webCLImageDescriptorObject.rowPitch * webCLImageDescriptorObject.height;
    var invalidByteLength = min - 1;
    data.byteLength = invalidByteLength;
    shouldThrowExceptionName("webCLContext.createImage(webcl.MEM_READ_WRITE, webCLImageDescriptor, data);", "INVALID_HOST_PTR");

    // FIXME :: To add test case to verify exception: IMAGE_FORMAT_NOT_SUPPORTED

} catch(e) {
    testFailed("WebCLContext :: createImage throws an exception : " + e.message);
}
notifyFinishedToHarness();
successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
